requirements.txt
Flask==2.0.3
PyPDF4==1.27.0
werkzeug
urllib3
requests
wheel